requirements.txt
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
attrs==20.2.0
autopep8==1.5.4
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.2.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
lxml==4.6.2
mccabe==0.6.1
packaging==20.8
Pillow==8.1.0
pycodestyle==2.6.0
pycparser==2.20
pylint==2.6.0
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
python-docx==0.8.10
selenium==3.141.0
six==1.15.0
soupsieve==2.1
toml==0.10.2
urllib3==1.26.3
webencodings==0.5.1
wrapt==1.12.1
